# db: Add index to speed up extension list query