--- a/migrations/frontend/1528395897_extensions_list_missing_index.down.sql
+++ b/migrations/frontend/1528395897_extensions_list_missing_index.down.sql
@@ -1,0 +1,3 @@
+BEGIN;
+DROP INDEX IF EXISTS registry_extension_releases_registry_extension_id_created_at;
+COMMIT;