--- a/migrations/frontend/1528395897_extensions_list_missing_index.up.sql
+++ b/migrations/frontend/1528395897_extensions_list_missing_index.up.sql
@@ -1,0 +1,1 @@
+CREATE INDEX CONCURRENTLY IF NOT EXISTS registry_extension_releases_registry_extension_id_created_at ON registry_extension_releases(registry_extension_id, created_at) WHERE deleted_at IS NULL;